usb: ehci: tegra: use suspend_noirq()/resume_noirq()
usb driver needs to be suspended late and resumed early even
before irqs are disabled/enabled. Without this change the following
two issues are seen during lp0 and lp1 states.

In lp0, when there is a usb wakeup event (by unplugging the usb device)
we get the following error:
	"tegra-ehci tegra-ehci.2:fatal error"
	"tegra-ehci tegra-ehci.2: HC died; cleaning up"
The above error comes because an irq is generated even before the
usb_resume was called.

A similar issue is seen in lp1 as well.

Bug 954564
Signed-off-by: Preetham Chandru R <pchandru@nvidia.com>
Change-Id: Id25fd2588ec034bd6aa54c17607e322f412adc5c
Reviewed-on: http://git-master/r/95778
Reviewed-by: Simone Willett <swillett@nvidia.com>
Tested-by: Simone Willett <swillett@nvidia.com>