misc: rfkill: rephrase gpio warning messages
The current BCM4329 rfkill driver assumes usage of 2 GPIOs
known as RST and SHUTDOWN and the driver makes a particular
GPIO mandatory. Some of the platforms does not define both
GPIOs, instead a single either RST/SHUTDOWN GPIO is used to
setup the chip.
Hence print a less alarming message if gpio is not found.

Bug 975333
Signed-off-by: Preetham Chandru r <pchandru@nvidia.com>

Change-Id: I229844d9d5dfc94ea00bb086d1091ac41bf97b99
Reviewed-on: http://git-master/r/106158
Tested-by: Preetham Chandru <pchandru@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Kiran Adduri <kadduri@nvidia.com>
Reviewed-by: Rakesh Goyal <rgoyal@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
GVS: Gerrit_Virtual_Submit