arm: tegra: grouper: Board files settings for LP1 0.95V CoreV Set the register values for each of the board files to keep the Core voltage to 0.95V in LP1. Using kai values at the moment but I have no idea if these will work for grouper.
This change is only for those platforms where LP1 is supported.
Enterprise and Kai and Grouper are the main platforms for this change. There
is no support for Cardhu for LP1 and so is left blank and the
feature will be skipped for Cardhu platforms, except for AP37.
AP37 with a PM269 board needs this change and so Cardhu board
file is updated with the values specific to AP37.

This change is part of the feature to set VCore to 0.95V
Refer to http://git-master/r/124135 for more details
Bug 1035684